chore: organize project structure with MVVM layers
