+++ lib/src/core/router/router.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
-import '../screens/home_screen.dart';
-import '../screens/setting_screen.dart';
+import '../../presentation/views/home/home.dart';
+import '../../presentation/views/setting/setting.dart';
 
 class AppRouter {
   static final GoRouter router = GoRouter(
