+++ lib/src/core/utils/constants.dart
@@ -1,0 +1,12 @@
+// lib/src/core/utils/constants.dart
+
+/// 앱 전역에서 사용하는 상수 값들을 정의하는 파일입니다.
+///
+/// 이 파일에는 런타임 중 변경되지 않는 값들을 모아둡니다. 예를 들어:
+///  - API 엔드포인트 URL
+///  - SharedPreferences 키 값
+///  - 기능 토글(Feature Flag) 이름
+///  - 열거형처럼 사용하는 문자열 또는 숫자 코드
+///
+/// 상수를 한곳에 모아두면 중복을 방지하고, 수정이 필요할 때
+/// 한 파일만 변경하면 되어 유지보수가 용이합니다.
