--- lib/src/models/model.dart
@@ -1,1 +1,0 @@
-// model.dart
