--- lib/src/services/service.dart
@@ -1,1 +1,0 @@
-// service.dart
